Paths Exception from System Property Value Containing Unix Start-Slash
This will require its own ticket to fix, since the system property
"javax.net.ssl.keyStore" is set in multiple locations and tests. For
now, in order to preserve code coverage, I filtered out the troublesome
initial slash that should not be on windows by detecting the colon using
a regex.